+++ src/main.js
@@ -1,32 +1,26 @@
-import { spawn } from 'child_process'
 import { env } from 'process'
 
 import getNode from 'get-node'
-import pEvent from 'p-event'
 
 import { validateInput } from './validate.js'
+import { spawnProcess } from './spawn.js'
 
 // Forwards `args` to another node instance of a specific `versionRange`
-const nve = async function(versionRange, args = [], opts = {}) {
-  validateInput(args)
+// eslint-disable-next-line max-params
+const nve = async function(versionRange, command, args = [], opts = {}) {
+  validateInput({ versionRange, command, args, opts })
 
-  // Download the Node.js binary
-  const progress = env.NVE_PROGRESS !== '0'
-  const { path: nodePath } = await getNode(versionRange, { progress })
+  const nodePath = await getNodePath(versionRange)
 
-  // Forward arguments to another node binary located at `nodePath`.
-  // We also forward standard streams.
-  const childProcess = spawn(nodePath, args, { ...DEFAULT_OPTIONS, ...opts })
-
-  const promise = getResult(childProcess)
-  return { childProcess, promise }
+  const childProcess = spawnProcess({ nodePath, command, args, opts })
+  return childProcess
 }
 
-const DEFAULT_OPTIONS = { stdio: 'inherit' }
-
-const getResult = async function(childProcess) {
-  const [code, signal] = await pEvent(childProcess, 'exit', { multiArgs: true })
-  return { code, signal }
+// Download the Node.js binary
+const getNodePath = async function(versionRange) {
+  const progress = env.NVE_PROGRESS !== '0'
+  const { path: nodePath } = await getNode(versionRange, { progress })
+  return nodePath
 }
 
 // We do not use `export default` because Babel transpiles it in a way that